se agregan nuevos cambios en materiaservice para id carrera

+++ src/main/java/com/ien/ienapp/entity/Inscripciones.java
@@ -39,7 +39,7 @@ public class Inscripciones {
     private PlanesEstudios planesEstudios;
 
     @Column(name = "nu_dni", nullable = false)
-    private Integer nuDni;
+    private BigInteger nuDni;
 
     @Column(name = "de_nombre", nullable = false)
     private String deNombre;
